Create validate_request.py
Make sure that all properties related to vector database given are sent
in the request. The properties required for each vectordb is different
and as such cannot be validated using the RequestModel. Here the
properties are checked considering the database selected. For example
connecting to pinecone requires an API key but connecting to chromadb
requires host and port only. Likewise different databases have different
properties which need to be validated.